# fix(bit-set/vector): counting set/flipped bits