@PhilippeMorier 2019-11-20
@Yomguithereal I realized that `Math.abs()` is not working in all cases. While doing #112 I removed a test which sets all the bits. Therefore, I didn't notice that `Math.abs()` doesn't actually work correctly :man_facepalming: :innocent:.